% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply two numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
Product of the two numbers
}
\description{
Multiply two numbers
}
\examples{
multiply(3, 4) # 12
}
