% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpkg2-package.R
\docType{package}
\name{testpkg2-package}
\alias{testpkg2}
\alias{testpkg2-package}
\title{testpkg2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramnathv/testpkg2}
  \item \url{https://ramnathv.github.io/testpkg2/}
  \item Report bugs at \url{https://github.com/ramnathv/testpkg2/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
